[release-branch.go1.17] cmd/link,internal/dwarf: decouple some common code with specific implement
To generate dwarf type info in compiler, I move some functions in cmd/link/internal/ld/dwarf.go to cmd/internal/dwarf, because they are independent of linker. No functional change now. They will be used in compiler later.

For #52209

Change-Id: I10c9bbaf530f84f3fe4d94687a9325b665a7a9ce

cmd/internal/dwarf: define interface for dwarf type info

Dwarf type info can be generated based on the abstract type. We can refact it to be independent of linker with "abstract type".

For #52209

Change-Id: I4d8fd3c19bcb9a22cdb037fb5f2b05e1a5907e23

[release-branch.go1.17] cmd/link: extract newtype1 function

No functional change, the logic of generating dwarf type info will be moved to cmd/internal/dwarf. So decouple it with the specific implement with linker. Not complete now.

Note: the dwsym will implement dwarf.Type interface temporarily, it will be removed at last.

For #52209

Change-Id: I2de0d9e47c78529e6021b5f74375f21bdb4ea070

[release-branch.go1.17] cmd/link: complete decoupling for newtype

Implement the interface dwarf.Type for dwsym, and use it in newtype1. Now, newtype1 is no related to any specific code with linker, it will moved to cmd/internal/dwarf next cl.

note:
Using dwsym as interface will cost a lot of time because of duplicate
function call. But in the future, all the dwarf type info will be generated in compiler, these will be removed.

For #52209

Change-Id: I30165d0bb7722924ed1f306a2807fabf3eb4c22f

cmd/link,cmd/internal/dwarf: move newtype to dwarf package

cmd/link/internal/ld/dwarf.go:/^func.newtype1 --> cmd/internal/dwarf/dwarf.go:/^func.NewType

For #52209

Change-Id: Ie95eff712418ee454a9c3887dcd7e7488a42b011

cmd/compile: implement dwarf.Type interface for types.Type

Implement cmd/internal/dwarf.Type interface fpr cmd/compile/internal/types.Type with a wrapper.

For #52209

Change-Id: I46ce4735bfcf900d9e17c14b6f69ea542eac0a8a

cmd/internal/dwarf: clarify symbol name and dwarfname

In compiler, the name for die must be exactly name, not the name with "". So distinguish them in NewDie

For #52209

Change-Id: I51320fd7a3f2fdba768535d514e00f3a707aaac9

cmd/internal/obj/dwarf: implement some base method for dwCtxt

Only generate type info for types defined in current compile unit. So only emit a reloc for subtype when DefGoType.
And add a stub type for eface and iface type, only need their name to be referenced.

For #52209

Change-Id: If723e0d9588c1e82cf1a129ac4c8ec7eb3b1d98a

cmd/compile: generate dwarf type info in compiler

Generate dwarf type info and dump it to objfile. It is not be used in linker now.
Some types need to synthesize and the prototypes are still generated in linker. I will resolve it in the future.

For #52209

Change-Id: Ie7c67197f930462ae4474f291e02d3009b6651bc

cmd/link: use the dwarf type info generated by compiler

It's time to begin using the dwarf type info generated by compiler in linker. Though dwinfo of many types is generated by compiler, but It is still necessary to check these types with fuction subtype, because the subtype of them is not defined.

For #52209

Change-Id: Ic12f4ed1f510454880d6ef9c8608b405c5404a92

cmd/link: refactor mkinternaltype and cleanup DefPtrTo

move mkinternaltype to internal/dwarf, then we can decouple synthesizeXXXtypes continue.

For #52209

Change-Id: I5d038c1f8f97fae36fb18d474ec78fb8269db584

cmd/link: decouple synthesizeXXXtypes with specific code in linker

They will be moved to internal/dwarf/dwarf.go in next CL, this CL is for we can review it more easily.

For #52209

Change-Id: I27831f65f998f5bc100182f598be25870ac916da

cmd/link,internal/dwarf: move synthesizeXXXtypes to internal/dwarf

Only a code move.

For #52209

Change-Id: I58f6c401898c2d3350fdf51d17ca059ec829dfa5

cmd/link: unify the key of prototypedies in linker and compiler

Remove the "type." prefix of the keys.The keys in compiler didn't prefix with "type."

For #52209

Change-Id: Ibd0b10f8a2b6ded1a157b87cea0ca397be6cf848

[release-branch.go1.17] cmd/compile: add ptototype of the types need synthesized to builtin

To synthesize dwarf type in compiler, the prototype of string, map, slice, channel is needed. It can't be easily got when compile the code out of runtime. So use add them to builtin so we can get the prototype information.

More detailed explanation:
When a map(eg: map[int]int) is used, we need more dwarf type info than we can see in language level. The pseudo type hash<int,int>, bucket<int,int>, []val<int> and []key<int> will be generated for debug more easy. They are "template" type, we need to fill them when a map is instantiated. And we can't define these type directly by exist go runtime type. So they should be added to buitin for filling the "template" when generating the dwarf type info of them.
So does others.

See https://github.com/golang/go/issues/52209#issuecomment-1099829590

For #52209

Change-Id: Ifcb1e15d2300323980b320673adb67df2ee07956

[release-branch.go1.17] cmd/compile: synthesize types in compiler

Now, we can synthesize types in compiler. And all the dwarf type sym can be kept by relocs. subtype in linker can be removed now. There is some undifined relocs to runtime type in dwarf type.
These relocs are added by compiler, it don't know the corresponding runtime type will be defined or not. So skip them in relocsym.

For #52209

Change-Id: Ief7683f8d9033ebf0f682f582e2fffde63587bf0

[release-branch.go1.17] cmd/link: remove the dwarf type info generation code in linker

Now, all the dwarf type info is generated in compiler, it is not necessary to keep these code in linker, remove it.

For #52209

Change-Id: I9cfe253bd4b4ea7d93d0c4cfbf6d309f566f4552

cmd/compile,internal/dwarf: cleanup some code

Expect no functional change.
1. Make some function in cmd/internal/dwarf unexport. No one call them out of the package.
2. Store some data and status on the dwctxt in internal/obj, instead of many global variables.
3. Remove the flag dwarfType I added before, use unified dwarf flag.

For #52209

Change-Id: I7683e53b6d7c7d015fc8b8bc778b72fe495ae2e4

cmd/compile: lazy load prototype die when synthesizing types

Do not need to prepare prototype die when compiling any compile unit.
Generate them only when we need.

For #52209

Change-Id: I4f8191ed7c3f9866580f574bfce6a939157d15f2

[release-branch.go1.17] cmd/compile,link: support generating dwarf info when -linkshared

Now, all the dwarf type can be generated by compiler, when using static link, every compiler unit generate its own dwarf type die. when using -linkshared, all the types current compile unit need can be collected. So we can generate corresponding dwarf type info for it and subtypes of it.
For synthesizetype, we also need to keep the type info of them when linkshared

For #52209

Change-Id: Ia70ba39708ce710c45d923d5b86dac91d7e9d616

cmd/compile: do not dump incorrect prototype when using -linkshared

the prototype must be dumped when runtime is compiled. when we compiling a non runtime package, it should keep minimal type info for prototypes, so, do not dump the incorrect prototype, such as runtime.g, runtime._type and soon. Replace them as a uintptr.

For #52209

Change-Id: Ib33ff92b5be50bf376bdf1931b726929a6186a71